\name{GBNcausal-package}
\alias{GBNcausal-package}
\alias{GBNcausal}
\docType{package}
\title{
Estimation of gaussian bayesian networks
}
\description{
This package was built around the main fucntion MCMC.GBN, which enables to estimate the causal effects (direct or total) of a gaussian bayesian network based on a set of data.

All the other functions are either used in this algorithm, or can be used to infer on the causal effects matrices. 
}
\details{
\tabular{ll}{
Package: \tab GBNcausal\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-08-19\cr
License: \tab What license is it under?\cr
}

}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

